Remove api client setup from app router
