+++ src/components/ServerConnections.js
@@ -1,19 +1,48 @@
 import { ConnectionManager, Credentials, ApiClient, Events } from 'jellyfin-apiclient';
+
 import { appHost } from './apphost';
 import Dashboard from '../utils/dashboard';
 import { setUserInfo } from '../scripts/settings/userSettings';
+import appSettings from '../scripts/settings/appSettings';
+
+const normalizeImageOptions = options => {
+    if (!options.quality && (options.maxWidth || options.width || options.maxHeight || options.height || options.fillWidth || options.fillHeight)) {
+        options.quality = 90;
+    }
+};
+
+const getMaxBandwidth = () => {
+    /* eslint-disable compat/compat */
+    if (navigator.connection) {
+        let max = navigator.connection.downlinkMax;
+        if (max && max > 0 && max < Number.POSITIVE_INFINITY) {
+            max /= 8;
+            max *= 1000000;
+            max *= 0.7;
+            return parseInt(max, 10);
+        }
+    }
+    /* eslint-enable compat/compat */
+
+    return null;
+};
 
 class ServerConnections extends ConnectionManager {
     constructor() {
         super(...arguments);
         this.localApiClient = null;
 
-        Events.on(this, 'localusersignedout', function (eventName, logoutInfo) {
+        Events.on(this, 'localusersignedout', (_e, logoutInfo) => {
             setUserInfo(null, null);
 
             if (window.NativeShell && typeof window.NativeShell.onLocalUserSignedOut === 'function') {
                 window.NativeShell.onLocalUserSignedOut(logoutInfo);
             }
+        });
+
+        Events.on(this, 'apiclientcreated', (_e, apiClient) => {
+            apiClient.getMaxBandwidth = getMaxBandwidth;
+            apiClient.normalizeImageOptions = normalizeImageOptions;
         });
     }
 
@@ -36,6 +65,13 @@ class ServerConnections extends ConnectionManager {
         this.setLocalApiClient(apiClient);
 
         console.debug('loaded ApiClient singleton');
+    }
+
+    connect(options) {
+        return super.connect({
+            enableAutoLogin: appSettings.enableAutoLogin(),
+            options
+        });
     }
 
     setLocalApiClient(apiClient) {
