+++ src/scripts/site.js
@@ -163,6 +163,10 @@ async function onAppReady() {
         import('../assets/css/ios.scss');
     }
 
+    Events.on(appHost, 'resume', () => {
+        ServerConnections.currentApiClient()?.ensureWebSocket();
+    });
+
     appRouter.start();
 
     if (!browser.tv && !browser.xboxOne && !browser.ps4) {
